agora menu da matrix tá quase ok
o menu da matrix não tá funcionando como o esperado, mas o código da biblioteca da matrix foi bem melhorado e modularizado, pois estava gerando erro com clock, include, tudo no mundo... Agora essa parte em si tá redondinha, só precisa ajeitar pra as funções serem chamadas bonitinhas.
