+++ botao_joystick.c
@@ -1,17 +1,17 @@
-#include "pico/bootrom.h"//configuração para teste do bootseel :P
+#include "pico/bootrom.h"  // Configuração para teste do bootsel
 #include "pico/stdlib.h"
 #include "botao_joystick.h"
 #include "menu.h"
-
-
-// Protótipos das funções
-void botao_callback(uint gpio, uint32_t eventos);
-void botao_init(uint8_t pino);
 
 // Variáveis para debounce
 static volatile uint32_t ultima_interrupcao_a = 0;
 static volatile uint32_t ultima_interrupcao_b = 0;
 static volatile uint32_t ultima_interrupcao_joystick = 0;
+
+// Variáveis globais para armazenar os valores de x e y do joystick
+extern int x;  // Declarado em outro arquivo (por exemplo, main.c)
+extern int y;  // Declarado em outro arquivo (por exemplo, main.c)
+extern volatile bool estado_inverter;  // Declarado em outro arquivo (por exemplo, main.c)
 
 // Função de callback para tratar interrupções dos botões
 void botao_callback(uint gpio, uint32_t eventos) {
@@ -22,6 +22,7 @@ void botao_callback(uint gpio, uint32_t eventos) {
             if (tempo_atual - ultima_interrupcao_b > DEBOUNCE_TIME) {
                 ultima_interrupcao_b = tempo_atual;
                 if (eventos & GPIO_IRQ_EDGE_FALL) {
+                    printf("Botão B pressionado\n");
                     reset_usb_boot(0, 0); // Entra no modo bootsel
                 }
             }
@@ -31,6 +32,7 @@ void botao_callback(uint gpio, uint32_t eventos) {
             if (tempo_atual - ultima_interrupcao_joystick > DEBOUNCE_TIME) {
                 ultima_interrupcao_joystick = tempo_atual;
                 if (eventos & GPIO_IRQ_EDGE_FALL) {
+                    printf("Botão do joystick pressionado\n");
                     estado_inverter = !estado_inverter; // Inverte o estado do joystick
                 }
             }
@@ -40,7 +42,8 @@ void botao_callback(uint gpio, uint32_t eventos) {
             if (tempo_atual - ultima_interrupcao_a > DEBOUNCE_TIME) {
                 ultima_interrupcao_a = tempo_atual;
                 if (eventos & GPIO_IRQ_EDGE_FALL) {
-                    menu_handle_button_a(); // Alterna entre o menu principal e o submenu
+                    printf("Botão A pressionado\n");
+                    menu_handle_button_a(x, y, estado_inverter);
                 }
             }
             break;
