+++ frames_3x5.c
@@ -1,0 +1,15 @@
+#include "frames_3x5.h"
+
+double f_01_fino[NUM_LEDS_FINO] = { // completamente apagado
+    0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0};
+
+double f_02_fino[NUM_LEDS_FINO] = { // todos acesos
+    1.0, 1.0, 1.0,
+    1.0, 1.0, 1.0,
+    1.0, 1.0, 1.0,
+    1.0, 1.0, 1.0,
+    1.0, 1.0, 1.0};
