+++ frames_3x5.h
@@ -5,18 +5,8 @@
 
 #define NUM_LEDS_FINO 15
 
-double f_01_fino[NUM_LEDS] = { // completamente apagado
-    0.0, 0.0, 0.0,
-    0.0, 0.0, 0.0,
-    0.0, 0.0, 0.0,
-    0.0, 0.0, 0.0,
-    0.0, 0.0, 0.0};
-
-double f_02_fino[NUM_LEDS] = { // todos acesos
-    1.0, 1.0, 1.0,
-    1.0, 1.0, 1.0,
-    1.0, 1.0, 1.0,
-    1.0, 1.0, 1.0,
-    1.0, 1.0, 1.0};
+// Declarações das variáveis como extern
+extern double f_01_fino[NUM_LEDS_FINO];
+extern double f_02_fino[NUM_LEDS_FINO];
 
 #endif
