+++ init_GPIO.h
@@ -3,7 +3,7 @@
 
 #include "pico/stdlib.h"
 
-#define DEBOUNCE_TIME 300
+#define DEBOUNCE_TIME 350
 #define OUT_PIN 7
 #define JSTICK 22
 #define BOTAO_A 5
