+++ letras_5x5.c
@@ -1,0 +1,263 @@
+#include "letras_5x5.h"
+
+// Definição das letras (A-Z)
+double f_A[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_B[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_C[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_D[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_E[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 1.0, 1.0, 1.0, 1.0
+};
+
+double f_F[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0
+};
+
+double f_G[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 1.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_H[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_I[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_J[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    1.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 1.0, 0.0, 0.0, 0.0
+};
+
+double f_K[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 1.0, 0.0,
+    1.0, 1.0, 1.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_L[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 1.0, 1.0, 1.0, 1.0
+};
+
+double f_M[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 0.0, 1.0, 1.0,
+    1.0, 0.0, 1.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_N[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 1.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_O[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_P[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0
+};
+
+double f_Q[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 1.0, 1.0,
+    0.0, 1.0, 1.0, 1.0, 1.0
+};
+
+double f_R[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_S[NUM_LEDS] = 
+{
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 1.0, 1.0, 1.0, 0.0,
+    0.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_T[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0
+};
+
+double f_U[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 1.0, 1.0, 0.0
+};
+
+double f_V[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 0.0, 1.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0
+};
+
+double f_W[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    1.0, 0.0, 1.0, 0.0, 1.0,
+    1.0, 1.0, 0.0, 1.0, 1.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_X[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 0.0, 1.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 1.0, 0.0, 1.0, 0.0,
+    1.0, 0.0, 0.0, 0.0, 1.0
+};
+
+double f_Y[NUM_LEDS] = 
+{
+    1.0, 0.0, 0.0, 0.0, 1.0,
+    0.0, 1.0, 0.0, 1.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0
+};
+
+double f_Z[NUM_LEDS] = 
+{
+    1.0, 1.0, 1.0, 1.0, 1.0,
+    0.0, 0.0, 0.0, 1.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 1.0, 0.0, 0.0, 0.0,
+    1.0, 1.0, 1.0, 1.0, 1.0
+};
+
+double f_exclamation[NUM_LEDS] = 
+{
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0,
+    0.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0
+};
+
+// Ponto
+double f_ponto[NUM_LEDS] = {
+    0.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 0.0, 0.0, 0.0, 0.0,
+    0.0, 0.0, 1.0, 0.0, 0.0
+};
+
+double *letras_5x5[] = {
+    f_A, f_B, f_C, f_D, f_E, f_F,
+    f_G, f_H, f_I, f_J, f_K, f_L,
+    f_M, f_N, f_O, f_P, f_Q, f_R,
+    f_S, f_T, f_U, f_V, f_W, f_X,
+    f_Y, f_Z, f_01, f_exclamation, f_ponto // Espaço, exclamação e ponto
+};
+
