+++ menu.c
@@ -4,16 +4,21 @@
 #include "pico/stdlib.h"
 #include "hardware/i2c.h"
 #include "direcao.h"  // Inclui o arquivo de direções
-#include "botao_joystick.h"
+#include "botao_joystick.h"  // Inclui a declaração de estado_inverter
 #include <string.h>
-
 
 // Definições
 #define NUM_OPTIONS_MAIN 2              // Número de opções no menu principal
-#define NUM_OPTIONS_SUB_1 4               // Número de opções no submenu
+#define NUM_OPTIONS_SUB_1 4             // Número de opções no submenu
 #define NUM_OPTIONS_SUB_2 3
 #define MENU_OPTION_SPACING 12          // Espaçamento vertical entre as opções
 #define MENU_CURSOR_POSITION 10         // Posição horizontal do cursor (seta ">")
+
+// Variáveis locais
+RGBColor cor = {200, 0, 50};
+int x = 0;
+int y = 0;
+static bool exibindo_funcao = false; // Controla se o sistema está exibindo uma função
 
 // Variáveis globais
 static const char *menu_options_main[NUM_OPTIONS_MAIN] = {"Joysticks", "Matrix de Leds"}; // Opções do menu principal
@@ -57,6 +62,26 @@ void menu_update(int x, int y, bool estado_inverter) {
     // Depuração: imprime os valores de x, y e estado_inverter
     printf("Valores de x: %d, y: %d, estado_inverter: %d\n", x, y, estado_inverter);
 
+    // Se estiver no modo de exibição, apenas atualiza as funções de exibição
+    if (exibindo_funcao) {
+        if (is_joysticks_selected) {
+            switch (selected_option) {
+                case 0: // Direcoes
+                    display_directions(x, y, estado_inverter);
+                    break;
+                case 1: // Porcentagens
+                    display_percentages(x, y);
+                    break;
+                case 2: // Gestos
+                    display_gestures(x, y);
+                    break;
+                default:
+                    break;
+            }
+        }
+        return; // Ignora a navegação pelo menu
+    }
+
     // Obtém a direção atual usando a função print_direction
     Direction dir = print_direction(x, y, estado_inverter);
 
@@ -97,16 +122,62 @@ void menu_update(int x, int y, bool estado_inverter) {
     // Redesenha o menu atual
     if (in_sub_menu) {
         if (is_joysticks_selected) {
-            menu_draw(menu_options_joysticks, NUM_OPTIONS_SUB_1);  // Desenha o submenu de Joysticks
+            menu_draw(menu_options_joysticks, NUM_OPTIONS_SUB_1);
         } else {
-            menu_draw(menu_options_matrix, NUM_OPTIONS_SUB_2);  // Desenha o submenu de Matrix de Leds
+            menu_draw(menu_options_matrix, NUM_OPTIONS_SUB_2);
         }
     } else {
-        menu_draw(menu_options_main, NUM_OPTIONS_MAIN);  // Desenha o menu principal
+        menu_draw(menu_options_main, NUM_OPTIONS_MAIN);
     }
 }
 
-void menu_handle_button_a() {
+
+// Função para exibir uma frase na matriz de LEDs
+void display_frase() {
+    printf("Exibindo frase...\n");
+    const char *frase = "OI";
+    RGBColor cor = {200, 0, 50};
+    double intensidade = 0.1;
+    int velocidade = 100;
+    int tamanho_matriz = 5;
+
+    exibir_frase_rolagem(frase, cor, pio, sm, intensidade, velocidade, tamanho_matriz);
+}
+
+void display_animacao() {
+    printf("Exibindo animação...\n");
+    RGBColor cor = {200, 0, 50};
+    double intensidade = 0.1;
+    int velocidade = 500;
+
+    while (true) {
+        exibir_frame(heart_01, cor, pio, sm, intensidade);
+        sleep_ms(velocidade);
+        exibir_frame(heart_02, cor, pio, sm, intensidade);
+        sleep_ms(velocidade);
+
+        if (!gpio_get(BOTAO_A)) {
+            printf("Botão A pressionado, saindo da animação...\n");
+            break;
+        }
+    }
+    exibir_frame(f_01, cor, pio, sm, intensidade);
+}
+
+
+// Função para tratar o botão A
+void menu_handle_button_a(int x, int y, bool estado_inverter) {
+    if (exibindo_funcao) {
+        // Se estiver exibindo uma função, volta ao submenu
+        exibindo_funcao = false;
+        if (is_joysticks_selected) {
+            menu_draw(menu_options_joysticks, NUM_OPTIONS_SUB_1); // Redesenha o submenu de joysticks
+        } else {
+            menu_draw(menu_options_matrix, NUM_OPTIONS_SUB_2);    // Redesenha o submenu da matriz
+        }
+        return;
+    }
+
     if (in_sub_menu) {
         // Verifica se a opção "Voltar" foi selecionada
         if ((is_joysticks_selected && strcmp(menu_options_joysticks[selected_option], "Voltar") == 0) ||
@@ -116,9 +187,36 @@ void menu_handle_button_a() {
             menu_draw(menu_options_main, NUM_OPTIONS_MAIN);
             printf("Voltando ao menu principal...\n");
         } else {
-            // Outras opções não fazem nada por enquanto
-            printf("Opção selecionada: %s\n", 
-                   is_joysticks_selected ? menu_options_joysticks[selected_option] : menu_options_matrix[selected_option]);
+            // Se estiver no submenu de Joysticks, entra no modo de exibição
+            if (is_joysticks_selected) {
+                exibindo_funcao = true; // Entra no modo de exibição
+                switch (selected_option) {
+                    case 0: // Direcoes
+                        display_directions(x, y, estado_inverter);
+                        break;
+                    case 1: // Porcentagens
+                        display_percentages(x, y);
+                        break;
+                    case 2: // Gestos
+                        display_gestures(x, y);
+                        break;
+                    default:
+                        break;
+                }
+            } else {
+                // Se estiver no submenu da Matriz, entra no modo de exibição
+                exibindo_funcao = true; // Entra no modo de exibição
+                switch (selected_option) {
+                    case 0: // Frase
+                        display_frase();
+                        break;
+                    case 1: // Animacao
+                        display_animacao();
+                        break;
+                    default:
+                        break;
+                }
+            }
         }
     } else {
         // Se estiver no menu principal, entra no submenu correspondente
@@ -147,6 +245,115 @@ void menu_handle_button_a() {
             printf("Índice de opção inválido: %d\n", selected_main_option);
         }
     }
+}
+
+// Função para exibir as direções no OLED
+void display_directions(int x, int y, bool estado_inverter) {
+    ssd1306_Fill(Black);  // Limpa o display
+
+    // Obtém a direção atual
+    Direction dir = print_direction(x, y, estado_inverter);
+
+    // Define a posição do cursor para exibir a direção
+    ssd1306_SetCursor(10, 10);
+    ssd1306_WriteString("Direcao: ", Font_7x10, White);
+
+    // Exibe a direção no OLED
+    switch (dir) {
+        case CIMA:
+            ssd1306_WriteString("Cima", Font_7x10, White);
+            break;
+        case BAIXO:
+            ssd1306_WriteString("Baixo", Font_7x10, White);
+            break;
+        case ESQUERDA:
+            ssd1306_WriteString("Esquerda", Font_7x10, White);
+            break;
+        case DIREITA:
+            ssd1306_WriteString("Direita", Font_7x10, White);
+            break;
+        case DIREITA_CIMA:
+            ssd1306_WriteString("Dir-Cima", Font_7x10, White);
+            break;
+        case DIREITA_BAIXO:
+            ssd1306_WriteString("Dir-Baixo", Font_7x10, White);
+            break;
+        case ESQUERDA_CIMA:
+            ssd1306_WriteString("Esq-Cima", Font_7x10, White);
+            break;
+        case ESQUERDA_BAIXO:
+            ssd1306_WriteString("Esq-Baixo", Font_7x10, White);
+            break;
+        default:
+            ssd1306_WriteString("Centralizado", Font_7x10, White);
+            break;
+    }
+
+    ssd1306_UpdateScreen();  // Atualiza o display
+}
+
+// Função para exibir as porcentagens no OLED
+void display_percentages(int x, int y) {
+    ssd1306_Fill(Black);  // Limpa o display
+
+    // Calcula as porcentagens
+    int x_processado = (x > MID + DEADZONE || x < MID - DEADZONE) ? x : MID;
+    int y_processado = (y > MID + DEADZONE || y < MID - DEADZONE) ? y : MID;
+
+    float porcentagem_x = ((x_processado - MID) / (float)MID) * 100;
+    float porcentagem_y = ((y_processado - MID) / (float)MID) * 100;
+
+    // Define a posição do cursor para exibir as porcentagens
+    ssd1306_SetCursor(10, 10);
+    ssd1306_WriteString("X: ", Font_7x10, White);
+    ssd1306_SetCursor(30, 10);
+    char x_str[10];
+    snprintf(x_str, sizeof(x_str), "%.2f%%", porcentagem_x);
+    ssd1306_WriteString(x_str, Font_7x10, White);
+
+    ssd1306_SetCursor(10, 20);
+    ssd1306_WriteString("Y: ", Font_7x10, White);
+    ssd1306_SetCursor(30, 20);
+    char y_str[10];
+    snprintf(y_str, sizeof(y_str), "%.2f%%", porcentagem_y);
+    ssd1306_WriteString(y_str, Font_7x10, White);
+
+    ssd1306_UpdateScreen();  // Atualiza o display
+}
+
+// Função para exibir os gestos no OLED
+void display_gestures(int x, int y) {
+    ssd1306_Fill(Black);  // Limpa o display
+
+    // Detecta o gesto
+    detectar_gesto(x, y);
+
+    // Define a posição do cursor para exibir o gesto
+    ssd1306_SetCursor(10, 10); // Posição X, Y (ajuste conforme necessário)
+    ssd1306_WriteString("Gesto:", Font_7x10, White);
+
+    // Exibe o gesto no OLED
+    Direction gesto = identificar_gesto();
+    switch (gesto) {
+        case GESTO_VOLTA_COMPLETA:
+            ssd1306_SetCursor(10, 25); // Posição X, Y (ajuste conforme necessário)
+            ssd1306_WriteString("Volta completa", Font_7x10, White);
+            break;
+        case GESTO_MEIA_LUA_DIREITA:
+            ssd1306_SetCursor(10, 25); // Posição X, Y (ajuste conforme necessário)
+            ssd1306_WriteString("Meia-Dir", Font_7x10, White);
+            break;
+        case GESTO_MEIA_LUA_ESQUERDA:
+            ssd1306_SetCursor(10, 25); // Posição X, Y (ajuste conforme necessário)
+            ssd1306_WriteString("Meia-lua Esq", Font_7x10, White);
+            break;
+        default:
+            ssd1306_SetCursor(10, 25); // Posição X, Y (ajuste conforme necessário)
+            ssd1306_WriteString("Nenhum", Font_7x10, White);
+            break;
+    }
+
+    ssd1306_UpdateScreen();  // Atualiza o display
 }
 
 // Função para obter a opção selecionada
